Create day-5 challenge code
Using the boilerplate, make a program that gets a list of countries from a website with their currency codes, then let the user choose a country and display the currency code of that country. This is part one of a bigger "country scrapping" project we will complete in the following days.