Create day-6 program
After getting the countries and their codes let the user choose two countries. Then let the user write an amount of currency they want to convert between country a and country b. Then send the two currency codes and the amounts to a URL like: https://transferwise.com/gb/currency-converter/gbp-to-usd-rate?amount=50
Use beautiful soup to scrape Transfer Wise to get the result of the conversion.